Se cambian los tipos de datos de id en las entidades Category, Event, Reservation y User de BigInteger a Long, y se actualiza la tabla de usuarios en la base de datos. Se añade un archivo SQL de importación con datos iniciales para categorías, usuarios y eventos.

+++ src/main/java/com/co/davivienda/ti/prueba/repositories/CategoryRepository.java
@@ -17,6 +17,6 @@ import com.co.davivienda.ti.prueba.entities.Category;
  * @since 2025/04/03
  */
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, BigInteger> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
 }
